requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
anthropic==0.8.1
python-dotenv==1.0.0
pydantic==2.5.3
openai==1.3.0
langchain==0.1.0
python-multipart==0.0.6
